feat(Device): add basic headset and controller tracking
This utilises the SteamVR plugin from the Unity Asset Store and
provides a basic `CameraRig` prefab that enables the headset tracking
in the game world along with controller tracking and creates the
controller models.

The actual prefab used is not the one from the SteamVR prefabs folder
as it seems to be missing the relevant controller elements but there
is a version of the prefab in the `SteamVR_TestThrow` scene that
contains all of the relevant elements so this is the one that has been
copied to the project's `Prefabs` directory.

A basic example scene has been added that demonstrates the `CameraRig`
headset and controller tracking.