feat(Controller): add button event listener script
A new Controller Events script has been added that listens for the
interactions with the VR controller and emits an event when a button
is pressed or released (also touched in the case of the touchpad).
It also holds the current state of each action in a public boolean
so it's possible to determine if a button is being continually held
down.

A new example scene has been added which demonstrates the Controller
events along with an example script which just logs the captured events
to the console.