feat(Pointer): add simple pointer for highlighting game items
A new `SimplePointer` script has been added that is attached to a
VR Controller in the `CameraRig` prefab and provides a basic laser
beam to be emitted from the end of the controller.

When the beam collides with another object in the game world, the
pointer emits an event along with a payload of the item it has
collided with and the distance the controller is from the object.

A new example scene has been added which demonstrates the Pointer as
well as demonstrates the events along with an example script which just
logs the pointer events to the console.

The `Unlit/Color` shader has been added to the project settings to
ensure it is included at build time otherwise the laser beam
won't be able to be created.