feat(Pointer): add new event for giving pointer end position
A new `PointerDestinationSet` event is now emitted when the pointer
beam is deactivated and the event payload now contains the world
position of the pointer tip which can be used to determine the position
in the world where the pointer was hitting.

This can be useful for determining where to teleport a player to by
listening to this new event and using the tip position variable to
move the player position to.

The accompanying example script has been updated to show this event.