feat(Controller): add custom button alias events
The Controller Events script now has three common action aliases:

  * Pointer Toggle
  * Interact Toggle
  * Menu Toggle

These aliases can be set to whichever button is desired to trigger them
and the other scripts now listen for the alias event rather than the
button event meaning their actions can be configured to a specific
button.

So now the Simple Pointer listens for the Pointer Toggle event and
therefore doesn't need to be tied to pressing the grip button.

And the Controller Interact script listens for the Interact Toggle
event rather than the Trigger so the interaction button can be
customised.