feat(Teleport): add teleportation to different object heights
A new Height Adjust Teleporter has been created that is extended from
the Basic Teleporter (which has now been updated to allow other
teleporters to inherit the Basic Teleport).

The Height Adjust Teleporter will check to see if the beam tip is on
top of the target object and if it is then the new teleport Y location
will be on top of that object, which enables climbing up objects in
the game world.

There is an additional flag of `playAreaFalling` that determines if
the player should fall down to the nearest floor if they step into
a part of the play area that is not on the object they are standing on.

If this flag is turned off then they will not fall down and can hover
in the air as their play area is fixed to the Y of the object they
currently standing on.

A new example scene has been added to show the new height adjust
teleporter working.