feat(Pointer): add bezier curve pointer
The Bezier Pointer always shows a cursor on the top of any object and
draws a bezier curve to the ground cursor so it is easier to see where
the teleport location will end up.

A helper file has been added that takes care of calculating the bezier
curve and plotting a collection of spheres through it to simulate the
pointer beam over a curve. This bezier curve code is heavily inspired
by the code and tutorial at:
`http://catlikecoding.com/unity/tutorials/curves-and-splines/`

It is possible to set a valid beam colour and an invalid beam colour
which will determine if the beam has found a valid floor to land on.

A new example has been added to show the new Bezier Pointer working
with teleportation.