chore(Pointer): add default target colours for pointer
As the shader is a transparent shader, the default colour when not
set is `Color.clear` which is also transparent which causes
confusion for users when using the script as the valid and invalid
target colours both have an alpha of zero.

This means that it's easy to set the colour of the target but forget
about setting the alpha so it seems as if the beams are broken but
in fact they are just completely transparent.

A simple solution is to just default the target colours with a
translucent colour so it's less confusing to new users.