feat(Interaction): add ability to trigger actions on touch events
Previously, the grab and use event could trigger actions in the
Interactable Object script (or scripts that extend it) but it was not
possible to override a method to deal with touching or stop touching.

The Interactable Object now supports overridable methods for
`StartTouching` and `StopTouching` and also sets a new flag of
`isTouching` on the object.

The Interact Touch class now also calls the Interactable Object touch
methods when the controller touches or stops touching the object.