fix(Teleport): ensure headset compensation works correctly
Previously, the headset compensation was not working and would
always set the teleport location to the centre of the play area.

This has now been fixed and a new `Headset Position Compensation`
option on the script has been added to toggle it.

Also, the basic teleporter can no longer traverse terrains as it
caused too many issues with an incorrect Y value on normal objects.