feat(Interaction): add ability to not need to hold buttons to interact
Previously, the grab button was required to be held down to continue
holding an object and the use button was also required to be held down
to ensure an object's Using state was active.

Now the Interactable Object script has two new parameters:

  * `holdButtonToGrab`
  * `holdButtonToUse`

If these parameters are checked then the previous action of needing to
hold the button down to perform the action is still used.

However, if they are unchecked then the user will have to press the
button once to initiate the action (e.g. press to grab) then they no
longer have to continually hold the button down to keep doing the
action (e.g. keep grabbing). When they want to stop doing the action
they press the button again (e.g. to stop grabbing).

This adds the benefit that multiple items can be set into their Using
state at any one time.

A new example scene shows this in action. The example scene has a red
barrel gun which demonstrates that setting the `holdButtonToUse`
parameter to unchecked doesn't always produce a good effect, in this
case the gun only fires on every other trigger press.