feat(Interaction): add example to show touchpad axis control
A new example has been added that contains an R/C car that is
controlled by moving a finger across the touchpad. Moving up and down
on the touchpad causese the car to drive forward and backward, moving
left and right causes the car to turn in that direction.

The trigger axis is also used to determine how high the car will jump
when the trigger is pressed.