chore(structure): remove inclusion of steamvr plugin
The SteamVR plugin is no longer included in this repository and needs
to be downloaded separately from the Unity Asset Store. This means that
upgrading the plugin will be independent of this toolkit keeping a
cleaner separation of concerns.

To ensure the separation can take place, a new script called
`SteamVR_ControllerActions` has been added that handles the hiding
of the controller models as this was originally added to the code
within the SteamVR plugin script `SteamVR_TrackerController`.

All of the example scenes that require to hide the controller now have
this new script added to the controllers so they function correctly.
The `SteamVR_InteractTouch` script will throw an error if it is
attached to a controller object that does not have the new
`SteamVR_ControllerActions` script attached to it.