feat(CameraRig): add ability to move camera rig with touchpad
The touchpad can now be used to move the play area in the direction
the headset is facing. It can also be used to strafe left and right
in the current facing direction.

The new `SteamVR_TouchpadWalking` script is added to the `[CameraRig]`
prefab and adds a box collider and a rigidbody to the camera rig to
handle collisions.

If the `SteamVR_HeadsetCollisionFade` script is also applied to the
`[CameraRig]->Camera` then on collision of the headset with a game
object, the user will have their position reset to their last good
known position. To achieve this a new `HeadsetCollisionDetect` event
has been added to the Headset Collision Fade script which the new
Touchpad Walking script listens for.

A big thanks to Reddit user `u/MisterDeum` for all the help with
testing this feature and coming up with ideas to make it a decent
feature to add to the toolkit.

A new example scene has been added to demonstrate this feature.