fix(Interaction): prevent spring joint when other controller attached
When an Interactable Object is grabbed, if that object has a joint on
it already causes the controller to be attached with a Spring Joint.
This causes an issue when another controller is grabbing the item
because it adds a Fixed Joint to the object causing the second grabbing
controller to have a Spring Joint added when grabbing from one hand
to the other.

The solution is to check to see if the existing joint is on a
controller and if it is then add a Fixed Joint instead.