fix(Interaction): prevent interacted object collision intersection
Because the Interactable Objects are fixed to the controller via a
joint and the controller can pass through any game object, this causes
collision resolution issues with objects being held. As the object
has a rigidbody and cannot enter the objects the controller can so
the physics engine struggles to resolve the collisions and the objects
flying around.

The solution is to apply a break force on the joint holding the object
to the controller, therefore if the controller tries to pull the
grabbed object into another object then eventually the physics engine
applies enough force on the grabbed object which can cause the joint
to break.

This means that if a grabbed object is pulled into another object then
the grabbed object is released just as if the user had banged their
hand and dropped the object.

The joint break limit can be configured on the Interactable Object by
changing the `Detatch Threshold` parameter.

Another parameter has been added to the Interactable Object to
determine the damper effect on the controller when interacting with
objects that already have a joint. The new `Joint Damper` parameter
will prevent oscillation when it is set to a high number. Also,
setting it to a high number will mean the object is needed to be
pulled more, which can simulate lifting heavy items.