feat(Pointer): add option to disable teleport on pointer
It may be required to have a controller pointer appear but never want
it to teleport, the use case would be one controller is used for
teleporting (perhaps with a bezier pointer) and the other pointer
uses a simple pointer but only for using Interactable Objects.

The new `Enable Teleport` flag can be set independently on each
controller.