feat(Pointer): add pointer activation delay option
It is now possible to provide a value to the `Activate Delay` parameter
on any pointer script that delays the pointer from being enabled again
for the number of seconds provided.

The seconds are roughly calculated based on the update routine having
60 frames per second.

This is useful to prevent teleporting immediately after the previous
teleport action.