fix(CameraRig): add movement deadzone to prevent walking glitch
The Touchpad Walking script had an issue where the speed variable
would not completely reset after releasing the touchpad meaning the
CameraRig would slide slowly across the floor as a value was still
present. This fix adds a deadzone margin to zero the speed variable
if it falls between this margin to prevent the sliding effect.