feat(Interaction): add rigidbody to controller option
There is now an option on the `InteractGrab` script that enables a
rigidbody on the controller object if the grab button is pressed when
the controller is not touching any Interactable Objects. This means
that the controller can be used to interact with other rigidbodies
to push them around or press them (e.g. a button).