feat(CameraRig): extract user rigidbody out of touchpad walking
The Touchpad Walking script no longer is responsible for generating the
collider and rigidbody around the user. This logic is now in it's own
script called `SteamVR_PlayerPresence` which is used to create the
collider and rigidbody and can be used independently with other use
cases for this feature. The Touchpad Walking script requires the
Player Presence script to be attached to the CameraRig as well for it
to function.

A new method for locating the controllers has been added rather than
looking for game objects with a controller script on. Now, the
controller event `device_connected` is listened to and on connection
the appropriate listeners are registered.

A new helper class called `DeviceFinder` has been added to handle
getting script reference points for connected devices.

The Basic Teleport script has also been updated to utilise this new
method of controller location to ensure a consistent solution.