fix(Teleporter): prevent falling through higher block on start
Previously, if the user started on a collidable object that was above
another collidable object and play space falling was checked, then on
start the user would fall through the object they were standing on
because on game initialisation the headset Y position (where the ray
is projected from) is the same as the play area Y position, so the ray
was being projected from underneath the collision point of where the
user was standing.

The fix is to only project the ray if the headset is above the
play area, then it will work after the headset is initialised
correctly.