fix(Controller): store connected device by object instead of index
The index of a connected device can change even on initialisation
meaning that the controller can be registered under different indexes
even when the game starts, meaning the listeners are registered
multiple times causing unwanted artifacts.

The solution is to track the connected controllers by the actual
object rather than the index of the device.