fix(CameraRig): get controllers for events from controller manager
Previously, the way scripts attached to the CameraRig found the
controllers to attach events to was to listen to the `device_connected`
event emitted by tracked objects and attempted to keep looping until
the listeners were attached.

This was flakey and did not work when transitioning scenes.

The CameraRig has a `SteamVR_ControllerManager` attached to it which
holds references to the two controller game objects so this is now
used to implicitly get the controller objects and ensure the
listeners are set up correctly.