feat(Interaction): add ability to auto grab an object
The new Object Auto Grab script can be attached to a controller and
be provided with a GameObject to automatically grab on game start.

It is also possible to clone the auto grabbed object (from a prefab or
existing scene object) which may be required for certain circumstances.

An example scene has been added to demonstrate this feature.