chore(Pointer): separate destination events from world pointer
The World Pointer class is no longer responsible for throwing the
destination events that is required by the Teleporter. The World
Pointer now inherits from a new Destination Marker class that throws
the destination events used by the teleporter.

This means other scripts could implement the destination events that
are not pointers, meaning other mechanisms for teleportation would
be possible.