chore(Version): bump version to 1.1.0
The latest version of the `README.md` has been included in PDF format
within the `SteamVR_Unity_Toolkit` directory to serve as an
instructions file required for the Unity Asset Store submission
process.

When ever a new version is about to be created, this `README.pdf` is
required to be updated.