feat(Presence): add toggle to reset position on headset collision
A new parameter has been added to the Player Presence script to
determine if the CameraRig position should be reset to a good known
position if the user's headset collides with another game object.

It is now possible with the `resetPositionOnCollision` parameter to
determine whether this action of position reset happens or not.