chore(Interaction): remove rotation snap grab snap type
The Rotation Snap is redundant as the Handle Snap is a far better
solution to the problem. The Rotation Snap Grab Snap Type and the
Snap To Position and Snap To Rotation parameters have been removed.

All of the example scenes that used Rotation Snap have now been
updated to use Handle Snap.