fix(Interaction): ensure press to grab/release works on quick grab
Previously, if the `Hold Button To Grab` parameter was unchecked and
an object was grabbed and dropped quickly, it didn't give the script
enough time to register the first event and set the touched object
correctly, meaning 3 button presses were required to drop an object
(or waiting about a second was required before dropping was allowed).

This has now been fixed by checking the grabbed object as a sanity
check if the touched object reports no object being held.