fix(Pointer): prevent pointer events when script is disabled
The pointer script would still throw the relevant events even if the
script component had been disabled at run time because the events had
already been set up. The fix is to check if the script is enabled and
only emit the events if it is.

The Basic Teleport has also been updated to find all Destination
Markers on an object and not just the first one so it's possible to
include multiple pointers on the same object and toggle them at
runtime.