feat(Interaction): allow custom rigidbody colliders for controller
It's now possible to pass a custom game object that contains a
rigidbody and the colliders for that rigidbody that will be used
to portray the shape of the controller. If no custom rigidbody is
provided then one is auto generated to match the shape of the
HTC Vive controller.

It is also possible to add a custom collider to the controller but if
no custom collider is found then a simple sphere collider is generated
itstead.