refactor(identification): determine player objects with script
Previously, player objects were identified by the name of the game
object. This was done by certain scripts would rename the game object
and add the string `PlayerObject` which would then be looked up by
other scripts.

This was not an ideal solution as it meant renaming game objects which
could create confusion and mess in the editor at runtime. It's cleaner
and more extensible to add a custom script to these objects and they
are identified by the existence of the script.

The `VRTK_PlayerObject` script also allows to hint at what type of
object it is from the following types:

  * CameraRig
  * Headset
  * Controller
  * Pointer