feat(Pointer): add option for custom pointer material
The World Pointer material can now be set as a parameter on any
pointer via the `Pointer Material` option in the inspector. If no
material is provided then a default material is loaded from the new
`Materials` directory.

This new `Materials` directory also provides the ability to load
required materials at runtime rather than trying to create them via
code using non-standard shaders. The approach of `Shader.Find` causes
build issues if the shader is not included in the
`Project Settings->Graphics->Shaders` option, but if there is a
material in the project that is using that shader then the build
process will automatically include the shader so there is no need to
manually update the settings.