fix(Presence): fix crash on object grab if no collider is present
The Player Presence script checks to see if a grabbed object has a
collider so it can ignore collisions with grabbed objects. However,
if the grabbed object doesn't have a collider but it's children do
then it would crash. This fix prevents that and checks for child
colliders to remove the physics collisions from.