fix(Pointer): ensure bezier pointer down beam is infinite
A previous change to the Bezier Pointer meant the projected down beam
was being limited to the Pointer Length parameter. The projected down
beam should always go as far down as it needs to find a ground point
so the max distance of the down beam has now been set to `infinity`.