fix(Teleport): ensure teleporting sets y position correctly
There are occasions when the pointer tip position would report being
above the target teleport location, or other times where the target
location wasn't being considered valid so the `[CameraRig]` y position
wasn't being updated leaving the player hanging in mid air.

This is because the logic to determine the destination y relied on
the target object having a defined center and height, which could
produce flakey results.

This has now been replaced and a ray is cast from the pointer tip
(with a slight height offset to ensure it avoids clipping with the
target surface). The ray is cast downwards and looks for the top of
the object the user should be standing on. This produces more
consistent results for both the bezier pointer and simple pointer
across many different surface types.

The terrain height check should also ensure negative height terrains
adhere to the correct destination height as well.