feat(DeviceFinder): add ability to get hand of controller
It's now possible to determine which hand the controller is for by
calling a new method `DeviceFinder.GetControllerHand(controllerObject);`
which can be useful for determining which controller is interacting
with an object. It can be used for example like so in the `StartUsing`
method of an interactable object:

`DeviceFinder.GetControllerHand(usingObject);`

This will determine which controller is being used to `Use` the obejct.