fix(Pointer): activate delay timer for teleport
Prevent multiple undesired teleport.

Even if the teleport activateDelayTimer is set, when users double-click
the teleport button, the teleport is activated twice.

Check the timer in VRTK_WorldPointer.SetPointerDestination() to prevent
this.

When the button is released let the beam be disabled immediately,
otherwise the beam cannot be activated again because of the timer and a
user feels like the button is not responding.