fix(Interaction): ensure ui hover events are correctly managed
Previously, the UI hover status was managed inside the custom Raycaster
but this was not working as expected as the hover array would get
cleared if multiple canvas items were present as the `RaycastAll` would
execute on all custom Raycasters.

The fix is to handle the hover state inline with each iteration over
the UI pointers in the custom EventSystem process. When a hover is
detected, then it is added to the hover array for processing further
down the chain.