fix(Pointer): switch off usage on pointer off
If pointerActivatesUseAction is true in VRTK_InteractableObject, after
pointing at the object with the pointer, the object continues to be used
also when the pointer is switched off.

Add a check in VRTK_WordPointer.TogglePointer () to switch off the
object usage if the pointer is switched off (add also a reference to the
object as private class member to keep track of the pointed object).