fix(Interaction): prevent infinite touch haptic loop
It could be possible to get into the situation where the touched
object could be not null and never equal the last touched object
meaning an infinite touch event would take place, causing the haptics
to infinitely loop.

The solution is to ensure when an object is actually touched it also
becomes the last touched object (which is logical). This should
ensure the infinite loop cannot occur.

It's also sensible to cancel any previous potential haptic events if
the touch is no longer occuring from touching something new.