fix(Interaction): react to enablement state of interactable_object
So far disabling an interactable_object at runtime did not have any
effect as it could still be touched and grabed. This patch checks the
enablement state explicitly.

A fix in the sample scene was needed to remove a boxcollider which
made it nearly impossible to showcase the feature that disables the
drawer contents.