feat(Interaction): enhance button control
The button control so far used constrained rigidbodies and a manual
check for distance. This turned out to be quite error prone when buttons
are rotated.

A better way to setup the button is with a limited configurable joint.
This way we don't need to do any manual checks and all constraint
calculations are done by the physics engine. This now also allows
rotated and more complex buttons.