feat(Interaction): enhanced drawer control
So far the drawer worked only with x and z axis. It now also detects
the y axis. In addition it supports handle-less drawers and the script
can be put onto the mesh directly without the need to specify the body.

Also a snapping mode has been added to compensate for minor physics
effects.