feat(Interaction): select objects by distance from attach point
Adds the multiTouchSelectionStyle option to VRTK_InteractTouch.
Allows user to specify how to disambiguate when the controller is over
multiple objects. The existing style selected the most recently touched
object. This commit also adds the ClosestOrigin style, which selects
the object with the closest origin to the controller attach point.