refactor(Structure): apply coding guidelines
An addition to the coding guidelines has been made to the
`CONTRIBUTING.md` file which states that all references must be in a
simplified state.

This new guideline has been applied to all of the existing code. Also
other guidelines that had been neglected to be applied have also been
applied.