fix(Interaction): ensure dropping objects on disable works
Previously, if the system dashboard was activated by pressing the
system menu button, then the `OnDisable` script would be called on
the Interactable Object, casuing the object to be dropped, which is
not the desired effect. When the dashboard is activated, the
controllers are disabled in Unity so now the force drop on disable
checks to see if the controllers are active before stopping the
interaction.

Also, when the object was a child of the controller, it would not
correctly unparent itself from the controller when it was disabled.
This has also been fixed.