feat(Pointer): add ability to manually toggle pointer beam
a new method has been added to the World Pointer that provides the
ability to manually toggle a pointer beam on or off without needing
the Pointer event being raised.

The `ToggleBeam(bool state)` method takes the state of whether the
beam should be `true` or `false`.

A new method has been added to the Device Finder to allow to find
a controller index by passing through a controller game object and
it will return the index in a uint.