fix(ControllerEvents): fix bug with SteamVR keyboard and uipointer
If SteamVR Keyboard is shown, SteamVR disables both controllers in the
scene. If the UIPointer is currently active (will happen if using the
UI, for example if input field is clicked), VRTK_ControllerEvents will
never change state  because controllers are disable. so UIPointer won't
act accordingly, This causes unwanted behaviours as even if no Pointer
is active, you can still interact with the UI.