fix(ControllerEvents): ensure all events are released when disabled
There is an issue that controller events such as button presses can
become stuck in the on state if the controller is disabled whilst a
button is being held down. For example, bringing up the SteamVR
keyboard overlay will disable the Unity controller models but then
when the overlay is closed the initial state of the controller's
pressed buttons is not updated so buttons are reported as being
pressed even though they are no longer being pressed.

This fix is to ensure that upon the controller being disabled that
all buttons that are active will have their released state emitted.
This will put them back in a fresh state when the controller is
re-enabled. However, if a button is being held down it will need
to be released and re-held down again for the event to register.