fix(example): prevent menu cube from detaching on activation
The example scene `019_ControllerInteractingWithPointer` has a menu
cube that is displayed when the app menu button is pressed. However,
due to some previous changes, the cube has it's parent reset due to
the `LoadPreviousState` method defaulting the parent to null.

The solution is to ensure `SaveCurrentState` is called so it always
knows about it's parent.