fix(Interaction): ensure grab state is reset for tracked objects
Minor fix for tracked objects, the grab state was not being reset when
"Hold Button To Grab" was disabled after dropping an item.
When grabbing it for a second time you have to hold the button
despite it being disabled. Moving the state reset into
InitUngrabbedObject() fixes the issue as it is called no
matter which path is used to drop an item.