fix(Interaction): wait for attach point before auto grabbing
The Object Auto Grab script could hit a race condition where the
`controllerAttachPoint` on the Interact Grab script hadn't been set
up in time and would cause the script to crash. This solution makes
the `Start()` method a coroutine that waits until the
`controllerAttachPoint` is available before proceeding to do the
auto grabbing of the object.