fix(Controller): ensure controller device index is valid
If the controllers are not ready or disabled then it's possible for
SteamVR to set the `controllerIndex` of a device to the max uint value
which essentially means it's invalid. If this happens then the
Controller Events script would crash as the device index was not valid.

This fix checks to see if the index is within a valid range before
proceeding to obtain the device.

fixes #360