fix(UIPointer): get pointer id from specified controller
Previously, the UIPointer would attempt to assign the pointer id
based on the provision that a Tracked Object script was present on
the same component as the UIPointer script. This meant it could
only be applied to the controllers or the headset by default.

The pointer id is now determined by the index of the controller
passed to the UI Pointer and therefore the UI pointer can be
attached to any game object that can also emit a world pointer.