feat(Controller): add more trigger states
The Controller Events script now has more states for the controller
trigger. Previously, it only emitted an event when the trigger
was pressed and released. But now it has 4 distinct states:

  * `Touched` - The trigger is begun to be squeezed.
  * `Pressed` - The trigger is squeezed about half way.
  * `Hairline` - The trigger is squeezed over the hairline threshold.
  * `Clicked` - The trigger is squeezed all the way until it clicks.

Because of these changes, the Controller Events inspector parameters
will be out of synchronization and will need manually updating when
applying this feature to an existing project (or the Controller Events
script will need resetting in the inspector to get the new default
button presses).