feat(UIPointer): add hold button to activate option
The UI Pointer can now also be toggled on and off with a button press
in the same way the World Pointer can. This means it works better
with a World Pointer that has the `Hold Button To Activate` parameter
unchecked as the world pointer will stay active until another button
press and now so will the UI pointer.